Moving to modular JS Structure
