+++ trip.html
@@ -86,7 +86,7 @@
 
     <!-- Your Config and Page Logic -->
     <script src="js/firebase-config.js"></script>
-    <script src="js/trip.js"></script>
+    <script type="module" src="js/trip.js"></script>
 
 </body>
